Move lib/slack/commands to lib/commands
