+++ lib/commands/help.js
@@ -1,4 +1,4 @@
-const { Help } = require('../../messages/flow');
+const { Help } = require('../messages/flow');
 
 module.exports = async (req, res) => {
   const { command } = res.locals;
