+++ lib/commands/list-subscriptions.js
@@ -1,6 +1,6 @@
 
-const SubscriptionList = require('../../messages/subscription-list');
-const { Help } = require('../../messages/flow');
+const SubscriptionList = require('../messages/subscription-list');
+const { Help } = require('../messages/flow');
 
 /**
  * Lists all subscriptions in a slack channel
