+++ lib/commands/signin.js
@@ -1,6 +1,6 @@
-const { SignIn } = require('../../messages/flow');
+const { SignIn } = require('../messages/flow');
 
-const SignedParams = require('../../signed-params');
+const SignedParams = require('../signed-params');
 
 // @todo should also be able to sign out (/github signout)
 module.exports = async (req, res) => {
