+++ lib/commands/subscribe.js
@@ -1,6 +1,6 @@
 const {
   Subscribed, NotFound, AlreadySubscribed, NotSubscribed, UpdatedSettings,
-} = require('../../messages/flow');
+} = require('../messages/flow');
 
 /**
  * Subscribes a slack channel to activity from an Organization or Repository
