+++ lib/slack/index.js
@@ -1,7 +1,7 @@
 const bodyParser = require('body-parser');
 
 const unfurl = require('./unfurl');
-const commands = require('./commands');
+const commands = require('../commands');
 const oauth = require('./oauth');
 const importer = require('./importer');
 const uninstall = require('./uninstall');
